usm: Move IsTLS to tls package
